requirements.txt
certifi==2025.1.31
cftime==1.6.4.post1
netCDF4==1.7.2
numpy==2.2.5
polars-lts-cpu==1.27.1
